[dhcpmon]: Saturate unhealthy episode counts
Use a 64-bit bounded health-window count and clamp the legacy int duration payload, preventing long-running reported or indeterminate episodes from overflowing.


Copilot-Session: 39f979be-d826-4d5c-949a-f20abb58bb83

+++ src/health_check.h
@@ -16,7 +16,7 @@ typedef struct
     dhcp_mon_status_t (*check_health)();           /** check function */
     void (*alert)(int duration);                   /** alert function when check failed */
     void (*log)(int duration);                     /** log function when check passed */
-    int count;                                     /** count in the number of unhealthy checks */
+    int64_t count;                                 /** count in the number of unhealthy checks */
     bool reported;                                 /** whether the current unhealthy episode was reported */
 } dhcp_mon_state_t;
